refactor resource page to conform to pattern.
